server restart when client has auth cookie is working
