+++ client/src/pages/index.js
@@ -8,8 +8,22 @@ import User from "./user";
 import VideoUpload from "./video_upload";
 import Navbar from "./Navbar";
 
+import { useLoadSession } from "../customHooks/useLoadSession";
+import Loading from "../components/Loading";
 
 export default function Pages() {
+  
+
+  React.useEffect(() => {
+    // Cleanup function: remove temp-session token from sessionStorage on unmount
+    return () => {
+      sessionStorage.removeItem("tempSessionToken");
+    };
+  }, []); // empty dependency array → cleanup on unmount
+
+  // This will handle fetching session from the server and saving temp-session from cookie
+  const sessionLoaded = useLoadSession();
+  if (!sessionLoaded) return <Loading />;
   return (
     <BrowserRouter>
       <Box
